--- lab-1/List/Sample-Test1.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29230.47
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Sample-Test1", "Sample-Test1.vcxproj", "{9E150E7B-2C49-4FA5-A23E-3705CD44D547}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9E150E7B-2C49-4FA5-A23E-3705CD44D547}.Debug|x64.ActiveCfg = Debug|x64
-		{9E150E7B-2C49-4FA5-A23E-3705CD44D547}.Debug|x64.Build.0 = Debug|x64
-		{9E150E7B-2C49-4FA5-A23E-3705CD44D547}.Debug|x86.ActiveCfg = Debug|Win32
-		{9E150E7B-2C49-4FA5-A23E-3705CD44D547}.Debug|x86.Build.0 = Debug|Win32
-		{9E150E7B-2C49-4FA5-A23E-3705CD44D547}.Release|x64.ActiveCfg = Release|x64
-		{9E150E7B-2C49-4FA5-A23E-3705CD44D547}.Release|x64.Build.0 = Release|x64
-		{9E150E7B-2C49-4FA5-A23E-3705CD44D547}.Release|x86.ActiveCfg = Release|Win32
-		{9E150E7B-2C49-4FA5-A23E-3705CD44D547}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D8A7A1EB-3F76-40F5-9FC9-AD7628CC620C}
-	EndGlobalSection
-EndGlobal
