--- lab-1/List/pch.cpp
@@ -1,6 +1,0 @@
-//
-// pch.cpp
-// Include the standard header and generate the precompiled header.
-//
-
-#include "pch.h"
